requirements.txt
python-dotenv==1.0.0
smartapi-python==1.2.6
pandas==2.0.3
numpy==1.24.3
tensorflow==2.12.0
transformers==4.30.2
torch==2.0.1
talisman==1.0.0
flask==2.3.2
schedule==1.1.0
requests==2.31.0
streamlit==1.29.0
plotly==5.18.0